+++ apps/backend/src/modules/routes/route.repository.ts
@@ -1,24 +1,49 @@
 import { SortingOrder } from "~/libs/enums/enums.js";
-import { type Repository } from "~/libs/types/types.js";
+import {
+	type Repository,
+	type TransactionOptions,
+} from "~/libs/types/types.js";
 
+import { type PlannedPathModel } from "../planned-paths/planned-path.model.js";
 import { type RouteFindAllOptions } from "./libs/types/types.js";
 import { RouteEntity } from "./route.entity.js";
 import { type RouteModel } from "./route.model.js";
 
 class RouteRepository implements Repository {
+	private plannedPathModel: typeof PlannedPathModel;
 	private routesModel: typeof RouteModel;
 
-	public constructor(routesModel: typeof RouteModel) {
+	public constructor(
+		routesModel: typeof RouteModel,
+		plannedPathModel: typeof PlannedPathModel,
+	) {
+		this.plannedPathModel = plannedPathModel;
 		this.routesModel = routesModel;
 	}
 
-	public async create(entity: RouteEntity): Promise<RouteEntity> {
+	public async create(
+		entity: RouteEntity,
+		options?: TransactionOptions,
+	): Promise<RouteEntity> {
 		const insertData = entity.toNewObject();
 
-		const result = await this.routesModel
-			.query()
+		let query = this.routesModel.query();
+
+		if (options?.transaction) {
+			query = query.transacting(options.transaction);
+		}
+
+		const result = await query
 			.insertGraph(insertData, { relate: ["pois"] })
-			.returning(["id", "name", "description"]);
+			.returning([
+				"id",
+				"name",
+				"description",
+				this.routesModel.raw("to_json(distance)::json as distance"),
+				this.routesModel.raw("to_json(duration)::json as duration"),
+				this.routesModel.raw("ST_AsGeoJSON(geometry)::json as geometry"),
+				"user_id",
+			]);
 
 		return RouteEntity.initialize(result);
 	}
@@ -42,7 +67,15 @@ class RouteRepository implements Repository {
 					builder.select("points_of_interest.id", "routes_to_pois.visit_order");
 				},
 			})
-			.select("routes.id", "routes.name", "routes.description");
+			.select([
+				"routes.id",
+				"routes.name",
+				"routes.description",
+				this.routesModel.raw("to_json(distance)::json as distance"),
+				this.routesModel.raw("to_json(duration)::json as duration"),
+				this.routesModel.raw("ST_AsGeoJSON(routes.geometry)::json as geometry"),
+				"routes.user_id",
+			]);
 
 		if (name) {
 			query.whereILike("routes.name", `%${name.trim()}%`);
@@ -61,11 +94,11 @@ class RouteRepository implements Repository {
 						`ST_Distance(
 						poi.location::geography,
 						ST_SetSRID(ST_MakePoint(?, ?), 4326)::geography
-						) as distance`,
+						) as distance_points`,
 						[longitude, latitude],
 					),
 				)
-				.orderBy("distance", SortingOrder.ASC);
+				.orderBy("distance_points", SortingOrder.ASC);
 		}
 
 		const routes = await query;
@@ -82,7 +115,15 @@ class RouteRepository implements Repository {
 					builder.select("points_of_interest.id", "routes_to_pois.visit_order");
 				},
 			})
-			.select("routes.id", "routes.name", "routes.description")
+			.select([
+				"routes.id",
+				"routes.name",
+				"routes.description",
+				this.routesModel.raw("to_json(distance)::json as distance"),
+				this.routesModel.raw("to_json(duration)::json as duration"),
+				this.routesModel.raw("ST_AsGeoJSON(routes.geometry)::json as geometry"),
+				"routes.user_id",
+			])
 			.where("routes.id", id)
 			.first();
 
@@ -107,7 +148,15 @@ class RouteRepository implements Repository {
 				},
 			})
 			.where({ id })
-			.returning(["id", "name", "description"])
+			.returning([
+				"id",
+				"name",
+				"description",
+				this.routesModel.raw("to_json(distance)::json as distance"),
+				this.routesModel.raw("to_json(duration)::json as duration"),
+				this.routesModel.raw("ST_AsGeoJSON(routes.geometry)::json as geometry"),
+				"routes.user_id",
+			])
 			.execute();
 
 		return updatedRoute ? RouteEntity.initialize(updatedRoute) : null;