+++ apps/frontend/src/index.tsx
@@ -17,6 +17,7 @@ import { NotFound } from "~/pages/not-found/not-found.jsx";
 import { Dashboard } from "./pages/dashboard/dashboard.js";
 import { Explore } from "./pages/explore/explore.js";
 import { Landing } from "./pages/landing/landing.jsx";
+import { PointsOfInterestDetails } from "./pages/points-of-interest-details/points-of-interest-details.js";
 import { Profile } from "./pages/profile/profile.js";
 import { RouteDetails } from "./pages/route-description/route-details.js";
 import { Routes } from "./pages/routes/routes.js";
@@ -30,6 +31,10 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 				routes={[
 					{
 						children: [
+							{
+								element: <PointsOfInterestDetails />,
+								path: AppRoute.POINTS_OF_INTEREST_$ID,
+							},
 							{
 								element: <Auth />,
 								path: AppRoute.SIGN_IN,