+++ apps/frontend/src/libs/modules/map-client/libs/types/types.ts
@@ -1,4 +1,3 @@
-export { type PointGeometry } from "@smartscapes/shared";
 export {
 	type ControlPosition,
 	type GeolocateControlOptions,