+++ apps/frontend/src/libs/modules/store/libs/types/root-reducer.type.ts
@@ -2,7 +2,10 @@ import { type reducer as appReducer } from "~/modules/app/app.js";
 import { type reducer as authReducer } from "~/modules/auth/auth.js";
 import { type reducer as exploreReducer } from "~/modules/explore/explore.js";
 import { type reducer as locationReducer } from "~/modules/location/location.js";
-import { type reducer as pointsOfInterestReducer } from "~/modules/points-of-interest/points-of-interest.js";
+import {
+	type pointOfInterestDetailsReducer,
+	type pointsOfInterestReducer,
+} from "~/modules/points-of-interest/points-of-interest.js";
 import { type reducer as routeReducer } from "~/modules/routes/routes.js";
 
 type RootReducer = {
@@ -10,6 +13,7 @@ type RootReducer = {
 	auth: ReturnType<typeof authReducer>;
 	explore: ReturnType<typeof exploreReducer>;
 	location: ReturnType<typeof locationReducer>;
+	pointOfInterestDetails: ReturnType<typeof pointOfInterestDetailsReducer>;
 	pointsOfInterest: ReturnType<typeof pointsOfInterestReducer>;
 	route: ReturnType<typeof routeReducer>;
 };