+++ apps/frontend/src/libs/modules/store/store.module.ts
@@ -15,7 +15,8 @@ import { reducer as exploreReducer } from "~/modules/explore/explore.js";
 import { reducer as locationReducer } from "~/modules/location/location.js";
 import {
 	pointOfInterestApi,
-	reducer as pointsOfInterestReducer,
+	pointOfInterestDetailsReducer,
+	pointsOfInterestReducer,
 } from "~/modules/points-of-interest/points-of-interest.js";
 import { routeApi, reducer as routeReducer } from "~/modules/routes/routes.js";
 import { userApi } from "~/modules/users/users.js";
@@ -58,6 +59,7 @@ class Store {
 				auth: authReducer,
 				explore: exploreReducer,
 				location: locationReducer,
+				pointOfInterestDetails: pointOfInterestDetailsReducer,
 				pointsOfInterest: pointsOfInterestReducer,
 				route: routeReducer,
 			},