+++ packages/shared/src/libs/types/location-geometry.type.ts
@@ -1,7 +1,8 @@
 import { type LocationType } from "../enums/enums.js";
+import { type Coordinates } from "./coordinates.type.js";
 
 type PointGeometry = {
-	coordinates: [number, number];
+	coordinates: Coordinates;
 	type: typeof LocationType.POINT;
 };